+++ Engine/ac/global_game.cpp
@@ -912,7 +912,7 @@ void scrWait(int nloops) {
 
     play.wait_counter = nloops;
     play.key_skip_wait = 0;
-    GameLoopUntilEvent(UNTIL_MOVEEND,(long)&play.wait_counter);
+    GameLoopUntilEvent(UNTIL_MOVEEND,&play.wait_counter);
 }
 
 int WaitKey(int nloops) {
@@ -921,7 +921,7 @@ int WaitKey(int nloops) {
 
     play.wait_counter = nloops;
     play.key_skip_wait = 1;
-    GameLoopUntilEvent(UNTIL_MOVEEND,(long)&play.wait_counter);
+    GameLoopUntilEvent(UNTIL_MOVEEND,&play.wait_counter);
     if (play.wait_counter < 0)
         return 1;
     return 0;
@@ -933,7 +933,7 @@ int WaitMouseKey(int nloops) {
 
     play.wait_counter = nloops;
     play.key_skip_wait = 3;
-    GameLoopUntilEvent(UNTIL_MOVEEND,(long)&play.wait_counter);
+    GameLoopUntilEvent(UNTIL_MOVEEND,&play.wait_counter);
     if (play.wait_counter < 0)
         return 1;
     return 0;
