+++ Engine/script/script.cpp
@@ -742,7 +742,7 @@ int run_interaction_commandlist (InteractionCommandList *nicl, int *timesrun, in
           MoveObject (IPARAM1, IPARAM2, IPARAM3, IPARAM4);
           // if they want to wait until finished, do so
           if (IPARAM5)
-              GameLoopUntilEvent(UNTIL_MOVEEND,(long)&objs[IPARAM1].moving);
+              GameLoopUntilEvent(UNTIL_MOVEEND,&objs[IPARAM1].moving);
           break;
       case 15: // Object Off
           ObjectOff (IPARAM1);
@@ -821,12 +821,12 @@ int run_interaction_commandlist (InteractionCommandList *nicl, int *timesrun, in
           break;
       case 34: // Run animation
           scAnimateCharacter(IPARAM1, IPARAM2, IPARAM3, 0);
-          GameLoopUntilEvent(UNTIL_SHORTIS0,(long)&game.chars[IPARAM1].animating);
+          GameLoopUntilEvent(UNTIL_SHORTIS0,&game.chars[IPARAM1].animating);
           break;
       case 35: // Quick animation
           SetCharacterView (IPARAM1, IPARAM2);
           scAnimateCharacter(IPARAM1, IPARAM3, IPARAM4, 0);
-          GameLoopUntilEvent(UNTIL_SHORTIS0,(long)&game.chars[IPARAM1].animating);
+          GameLoopUntilEvent(UNTIL_SHORTIS0,&game.chars[IPARAM1].animating);
           ReleaseCharacterView (IPARAM1);
           break;
       case 36: // Set idle animation
